requirements.txt
langchain
langchain-huggingface
pdfplumber
python-dotenv
unstructured
langchain-groq
streamlit
langchain-chroma
googleapis-common-protos
sentence-transformers